Move utils into shared folder
