+++ src/Sensors/TouchSensor/TouchSensor.js
@@ -1,5 +1,5 @@
+import {closest} from 'shared/Utils';
 import Sensor from './../Sensor';
-import {closest} from './../../Utils';
 
 import {
   DragStartSensorEvent,
